backend/requirements_py314.txt
# Python 3.14 compatible requirements
Django==5.0.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
psycopg==3.2.3
Pillow==10.4.0
gunicorn==23.0.0
whitenoise==6.8.2
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
cryptography==41.0.7
requests==2.31.0
